EIS: Constrained destination field to clause of target.
If the target is a feature, destination can now only be a pre or
postcondition.
If the target is a class, destination can only be an invariant.
Clauses are extracted from the AST and only one destination can for
now be associated to one source − in a future version, several
destinations should be possible.