EIS: Graphical and textual representation of destinations are now synchronized.
When you modify destination item in text, the entry is modified and the
graphical is synchronized.
When you modify destination item in graphical version, the note is also
modified.

If you change in text and put a bad entry (non-existing assertion), when
you load the interface, the destination is removed.